Re-organize folder structure and rename files to follow some convention. (#17)
/public is now /client and there is a /public subfolder to be used for
index.html and app_bundle.js. Actual code, including components, reducers, etc are in
/client/src. Server logic goes in /server/src. Test specs retain the same structure in /test.
File names are all lowercase with the exception of React components which are captialized. Any styles or utils supplemental to a component should have the component name + _ + 'styles' or
'util' for example. Additionally fixed all references so the app and specs work as before.