Connect client infrastructure (#42)
* Begin adding redux middleware.

* Configure store correctly to use initial state. Convert initial state to an Immutable JS object. Configure all reducer functions to set an initial state
to the equivalent property in the state tree. Save redux-immutable to package.json so that combineReducers will accept Immutable objects instead of plain JS. Add Provider
in Main.jsx so that the Redux store connects to our React app. Re-write Header.jsx to have both a stateless presentational component in addition to a container exported as default.
Additionally configure Header with mapStateToProps and mapDispatchToProps so it can access state properties and action creators from its props. Finally, modify a couple property names
in state.jsx.